Parametrized ObjId by the pointed-to object type.
This is equivalent to a static type tag and enables some additional
compiler type checks that are useful when writing Delta classes.

Existing code should be migrated so that every use of ObjId is parametrized
by a type. Also, we should investigate an AspectJ check that make sure
all slave objects storage/retrieval is done through the appropriate 
SlaveUpdater interface.




git-svn-id: file:///home/hoijui/src/IsaViz/tmp2/zvtm.svn/code@2634 74e165f3-c697-40a6-91af-552c8936d7fe